refactor: concentrate list handling fts in two files
