+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gwolf <gwolf@student.42vienna.com>         +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/07/28 13:03:05 by gwolf             #+#    #+#              #
-#    Updated: 2023/08/05 13:21:08 by gwolf            ###   ########.fr        #
+#    Updated: 2023/08/05 17:00:28 by gwolf            ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -93,6 +93,8 @@ SRC :=	utils_memory.c \
 		expand_handler.c \
 		handle_input_expand.c \
 		handle_input.c \
+		handle_input_tkn_list.c \
+		handle_input_tkn_list2.c \
 		parser.c \
 		parser_utils.c \
 		parser_fill_cmd.c \
